renaming and splitting of ham_tomography_utils.py file
